requirements.txt
flask
flask-script
flask-migrate
flask-redis
flask-sqlalchemy
qiniu
mysql-python
forgerypy
